% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCanalysis.R
\name{ROCanalysis}
\alias{ROCanalysis}
\title{Receiver Operating Characteristic Analysis}
\usage{
ROCanalysis(probList, Outcome, mat = NULL, ROCcurve = TRUE,
  AccCurve = TRUE)
}
\arguments{
\item{probList}{A vector of classification probabilities.}

\item{Outcome}{A vector of the responses (can either be factors or 0/1).}

\item{mat}{Option to include a confusion matrix when its value is in the interval [0, 1]. Defaults to NULL.}

\item{ROCcurve}{Creates an ROC curve with ggplot2. This includes the AUC. Defaults to TRUE.}

\item{ACCcurve}{Creates an optimal cutoff point graph. Defaults to TRUE.}
}
\description{
This function will create an ROC curve to visually demonstrate the performance of a logistic model.
}
\examples{
library(ISLR)
data("Weekly")

set.seed(1439)
train <- sample(nrow(Weekly), floor(nrow(Weekly)) * 0.75)
test.set <- Weekly[-train,]
#'
attach(Weekly)
log.model <- glm(Direction ~ Lag2, data = Weekly, subset = train, family = binomial)

predicted <- predict.glm(log.model, newdata = test.set, type = "response")

roc.log <- ROCanalysis(predicted, test.set$Direction, mat = 0.5)
roc.log$roccurve
roc.log$acccurve
roc.log$confusionMatrix
}
\keyword{ROC}
\keyword{classification}
\keyword{curve}
